nagbot: Remind people when it's time for standup
This is a quick hack with hard-coded time. Will make it configurable
later. Also suggest having a good weekend on Friday afternoon.

The code should work with multiple channels, but actually it's
untested since we currently use a single channel. I'm assuming we
will support multiple channels in future, hence channels is an
array.